Update MergedCosmicReader for new Cosmic releases
New Cosmic releases use a different column name for the Cosmic ID.
This is now fixed.

The cancer type count and cancer site count were based on study id.
This does not make sense, because we want to have the number of
tumours where a mutation is found, not the number of studies.
For now, I just replaced the study id column tag with the name of the
tumour id column. Addtional refactoring would be needed to rename
all references to _study_ with _tumour_.